add env example configure webpack for bundling move most code to src delete console logs
