+++ env.example
@@ -1,0 +1,4 @@
+DB
+SESS_NAME
+SESS_SECRET
+SESS_LIFETIME
