+++ src/common/helper.js
