+++ src/routes/users/controller.js
@@ -1,4 +1,4 @@
-import bcrypt from 'bcrypt';
+import bcrypt from 'bcryptjs';
 import User from '../../models/User';
 import Exam from '../../models/Exam';
