+++ webpack.config.js
@@ -1,0 +1,13 @@
+const path = require('path');
+const nodeExternals = require('webpack-node-externals');
+
+module.exports = {
+  mode: 'production',
+  target: 'node',
+  entry: './server.js',
+  output: {
+    path: path.resolve(__dirname, 'dist'),
+    filename: 'main.js',
+  },
+  externals: [nodeExternals()],
+};
